feat(backend): change type of field 'salary' in employees entity
change type of field 'salary' from integer to decimal in database, it was made to handle the storing of a financial value in a proper way

+++ backend/src/graphql/modules/Employees/schema.ts
@@ -40,7 +40,7 @@ export class AddEmployeeArgs {
   @IsMobilePhone('any')
   phone?: string;
 
-  @Field(() => Int)
+  @Field(() => Float)
   @Min(0)
   salary: number;
 
@@ -61,7 +61,7 @@ export class EditEmployeeArgs {
   @IsMobilePhone('any')
   phone?: string;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   @Min(0)
   salary: number;
 
@@ -180,7 +180,7 @@ export class Employee {
   @Field({ nullable: true })
   phone?: string;
 
-  @Field(() => Int)
+  @Field(() => Float)
   salary: number;
 
   @Field()
